requirements.txt
flask~=2.1.2
pyyaml>=5.4.1
waitress~=2.1.2
discord>=1.0.1
opylib>=0.7.0